backend/requirements.txt
Django>=3.2
djangorestframework
pandas
reportlab
django-cors-headers
